requirements.txt
werkzeug
soundfile
ffmpeg-python
numpy
torch
openai-whisper
tiktoken
pipdeptree 
setuptools_rust
ffmpeg 
# numba
# pyaudio
fastapi 
uvicorn>=0.27.1 
jinja2
python-multipart
